requirements.txt
numpy
pandas
finta
requests
python-telegram-bot